feat: 중복된 메뉴가 있으면 주문할 수 없다

+++ src/main/java/kitchenpos/domain/OrderLineItems.java
@@ -40,7 +40,7 @@ public class OrderLineItems {
         return Collections.unmodifiableList(elements);
     }
 
-    public void validateDuplicateMenu(int menuCount) {
+    public void validateDuplicateMenu(long menuCount) {
         if (elements.size() != menuCount) {
             throw new IllegalArgumentException("중복된 메뉴가 있습니다.");
         }
